requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
yfinance>=0.2.18
plotly>=5.15.0
ta>=0.10.2
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=11.0.0
schedule>=1.2.0